+++ dtaianomaly/preprocessing/__init__.py
@@ -8,12 +8,13 @@ Custom preprocessors can be implemented by extending the base :py:class:`~dtaian
 from .Preprocessor import Preprocessor, check_preprocessing_inputs, Identity
 from .ChainedPreprocessor import ChainedPreprocessor
 from .MinMaxScaler import MinMaxScaler
-from .ZNormalizer import ZNormalizer
+from .StandardScaler import StandardScaler
 from .MovingAverage import MovingAverage
 from .ExponentialMovingAverage import ExponentialMovingAverage
 from .UnderSampler import SamplingRateUnderSampler, NbSamplesUnderSampler
 from .Differencing import Differencing
 from .PiecewiseAggregateApproximation import PiecewiseAggregateApproximation
+from .RobustScaler import RobustScaler
 
 __all__ = [
     'Preprocessor',
@@ -21,11 +22,12 @@ __all__ = [
     'Identity',
     'ChainedPreprocessor',
     'MinMaxScaler',
-    'ZNormalizer',
+    'StandardScaler',
     'MovingAverage',
     'ExponentialMovingAverage',
     'SamplingRateUnderSampler',
     'NbSamplesUnderSampler',
     'Differencing',
-    'PiecewiseAggregateApproximation'
+    'PiecewiseAggregateApproximation',
+    'RobustScaler'
 ]