+++ tests/pipeline/test_Pipeline.py
@@ -1,6 +1,6 @@
 import pytest
 
-from dtaianomaly.preprocessing import Identity, ZNormalizer, ChainedPreprocessor
+from dtaianomaly.preprocessing import Identity, StandardScaler, ChainedPreprocessor
 from dtaianomaly.anomaly_detection import IsolationForest
 
 from dtaianomaly.pipeline import Pipeline
@@ -9,10 +9,10 @@ from dtaianomaly.pipeline import Pipeline
 class TestPipeline:
 
     def test_initialization(self):
-        Pipeline(ZNormalizer(), IsolationForest(15))
+        Pipeline(StandardScaler(), IsolationForest(15))
 
     def test_list_of_preprocessors(self):
-        Pipeline([ZNormalizer(), Identity()], IsolationForest(15))
+        Pipeline([StandardScaler(), Identity()], IsolationForest(15))
 
     def test_no_preprocessors(self):
         with pytest.raises(ValueError):
@@ -24,15 +24,15 @@ class TestPipeline:
 
     def test_invalid_list(self):
         with pytest.raises(TypeError):
-            Pipeline([ZNormalizer(), 'bonkers'], IsolationForest(15))
+            Pipeline([StandardScaler(), 'bonkers'], IsolationForest(15))
 
     def test_invalid_detector(self):
         with pytest.raises(TypeError):
-            Pipeline(ZNormalizer(), 'IsolationForest')
+            Pipeline(StandardScaler(), 'IsolationForest')
 
     def test_str(self):
-        assert str(Pipeline(ZNormalizer(), IsolationForest(15, 3))) == 'ZNormalizer()->IsolationForest(window_size=15,stride=3)'
-        assert (str(Pipeline(ChainedPreprocessor(Identity(), ZNormalizer()), IsolationForest(15, 3)))
-                == 'Identity()->ZNormalizer()->IsolationForest(window_size=15,stride=3)')
+        assert str(Pipeline(StandardScaler(), IsolationForest(15, 3))) == 'StandardScaler()->IsolationForest(window_size=15,stride=3)'
+        assert (str(Pipeline(ChainedPreprocessor(Identity(), StandardScaler()), IsolationForest(15, 3)))
+                == 'Identity()->StandardScaler()->IsolationForest(window_size=15,stride=3)')