+++ tests/preprocessing/test_Preprocessor.py
@@ -1,8 +1,7 @@
 import pytest
 import numpy as np
-import copy
 
-from dtaianomaly.preprocessing import Identity, MinMaxScaler, ZNormalizer, MovingAverage, ExponentialMovingAverage, ChainedPreprocessor, NbSamplesUnderSampler, SamplingRateUnderSampler, check_preprocessing_inputs
+from dtaianomaly.preprocessing import Identity, check_preprocessing_inputs
 
 
 class TestCheckPreprocessingInputs:
@@ -61,69 +60,6 @@ class TestCheckPreprocessingInputs:
             np.array([1, 2, 3, 4, 5]),
             None
         )
-
-
-@pytest.mark.parametrize('preprocessor', [
-    Identity(),
-    MinMaxScaler(),
-    ZNormalizer(),
-    MovingAverage(5),
-    ExponentialMovingAverage(0.5),
-    NbSamplesUnderSampler(100),
-    SamplingRateUnderSampler(5),
-    ChainedPreprocessor([Identity(), MinMaxScaler()])
-])
-class TestPreprocessors:
-
-    def test_invalid_input_fit(self, preprocessor):
-        with pytest.raises(ValueError):
-            preprocessor.fit(
-                np.array(['1', '2', '3', '4', '5']),
-                np.array([0, 0, 1, 0, 1])
-            )
-
-    def test_invalid_input_transform(self, preprocessor):
-        with pytest.raises(ValueError):
-            preprocessor.transform(
-                np.array(['1', '2', '3', '4', '5']),
-                np.array([0, 0, 1, 0, 1])
-            )
-
-    def test_invalid_input_fit_transform(self, preprocessor):
-        with pytest.raises(ValueError):
-            preprocessor.fit_transform(
-                np.array(['1', '2', '3', '4', '5']),
-                np.array([0, 0, 1, 0, 1])
-            )
-
-    def test_univariate(self, preprocessor, univariate_time_series):
-        preprocessor.fit_transform(univariate_time_series)
-
-    def test_multivariate(self, preprocessor, multivariate_time_series):
-        preprocessor.fit_transform(multivariate_time_series)
-
-    def test_fit(self, preprocessor, univariate_time_series):
-        preprocessor_copy = copy.deepcopy(preprocessor)
-        preprocessor.fit(univariate_time_series)
-        preprocessor_copy._fit(univariate_time_series)
-        if isinstance(preprocessor, ChainedPreprocessor):
-            for i in range(len(preprocessor.base_preprocessors)):
-                assert preprocessor_copy.base_preprocessors[i].__dict__ == preprocessor.base_preprocessors[i].__dict__
-        else:
-            assert preprocessor_copy.__dict__ == preprocessor.__dict__
-
-    def test_transform(self, preprocessor, univariate_time_series):
-        preprocessor.fit(univariate_time_series)
-        preprocessor_copy = copy.deepcopy(preprocessor)
-        x, _ = preprocessor.transform(univariate_time_series)
-        x_, _ = preprocessor_copy._transform(univariate_time_series)
-        assert np.array_equal(x, x_)
-
-    def test_fit_transform_different_time_series(self, preprocessor, univariate_time_series):
-        split = int(univariate_time_series.shape[0] / 2)
-        x_fit = univariate_time_series[:split]
-        x_transform = univariate_time_series[split:]
-        preprocessor.fit(x_fit).transform(x_transform)
 
 
 class TestIdentity: