+++ docs/additional_information/changelog.rst
@@ -12,6 +12,7 @@ Added
 - Implemented ``ClusterBasedLocalOutlierFactor`` (CBLOF) anomaly detector.
 - Implemented ``KMeansAnomalyDetector`` anomaly detector.
 - Implemented ``CopulaBasedOutlierDetector`` (COPOD) anomaly detector.
+- Implemented ``RobustScaler`` preprocessor.
 
 Changed
 ^^^^^^^
@@ -25,6 +26,7 @@ Changed
 
 Fixed
 ^^^^^
+- Renamed ``ZNormalizer`` to ``StandardScaler``, to make it align with the Sklearn declaration.
 
 
 [0.2.3] - 2024-12-02